WIP Metadata Article Search
Fix order of metadata/article name.
Modify buildQuery to be similar like springers: Build the search from
the manually given search terms.
Add queryUrl and queryHeaders as replacement for buildQuery
Add support for the metadata/article collection in translateQuery
Errors are not printed in utils.requestErrorHandling anymore but in
callAPI by adding the string representation of the exception.